+++ test/unit/start.spec.js
@@ -10,9 +10,9 @@ Unless required by applicable law or agreed to in writing, software distributed 
 */
 
 const sinon = require("sinon");
+const AdminApi = require("@mocks-server/plugin-admin-api");
 
 const CoreMocks = require("./Core.mocks.js");
-const AdminApi = require("../../lib/api/Api");
 const InquirerCli = require("../../lib/cli/Cli");
 
 const { start } = require("../../lib/start");